fix: fix shares reporting
